+++ src/common/components/markdownViewer/syncScroll/index.ts
@@ -1,1 +1,5 @@
-export { SOURCE_LINE_CLASSNAME, SOURCE_LINE_ATTRIBUTE } from './syncScroll';
+export {
+  SOURCE_LINE_CLASSNAME,
+  SOURCE_LINE_ATTRIBUTE,
+  getRenderedElementsForSourceLine,
+} from './synchronizeScroll';
